XMLHttpRequest/open-method-bogus.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() - bogus methods</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      function method(method) {
        test(function() {
          var client = new XMLHttpRequest()
          assert_throws("SYNTAX_ERR", function() { client.open(method, "...") })
        }, document.title + " (" + method + ")")
      }
      method("")
      method(">")
      method(" GET")
      method("G T")
    </script>
  </body>
</html>
